move code to subfolder and add WIP, pipeline
WIP is just that, ideas that are still too new to produce any solid analysis.

Pipeline is data that could conceivably be processed / updated directly to Google using GitHub Actions, but currently is aggregated using a manual execution of the scripts in code/.

The data in pipeline/ksde is enrollment data, broken out by grade, gender, and economic special sub-group. It's an aggregation of those categories across the district.